+++ src/boilerplate.rs
@@ -1,0 +1,51 @@
+use bevy::{app::PluginGroupBuilder, prelude::*};
+
+// Function to setup the app
+pub fn default_plugins() -> PluginGroupBuilder {
+    DefaultPlugins.set (
+        WindowPlugin {
+            primary_window: Some(Window {
+                title: "Bevypunk".into(),
+                mode: bevy::window::WindowMode::BorderlessFullscreen,
+                ..default()
+            }),
+            ..default()
+        }
+    )
+}
+
+// A struct to differentiate between other types of ui (required by Lunex)
+#[derive(Component, Debug, Default, Clone, PartialEq)]
+pub struct MenuUi;
+
+
+// Load all assets at startup for faster loading during runtime
+#[derive(Resource)]
+pub struct AssetCache {
+    pub font: Handle<Font>,
+    pub font_bold: Handle<Font>,
+    pub button: Handle<Image>,
+
+    pub switch_base: Handle<Image>,
+    pub switch_head: Handle<Image>,
+
+    pub main_background: Handle<Image>,
+    pub main_board: Handle<Image>,
+    pub main_logo: Handle<Image>,
+    pub settings_background: Handle<Image>,
+}
+pub fn presetup(mut commands: Commands, asset_server: Res<AssetServer>) {
+    commands.insert_resource(AssetCache {
+        font: asset_server.load("fonts/rajdhani/Rajdhani-Medium.ttf"),
+        font_bold: asset_server.load("fonts/rajdhani/Rajdhani-Bold.ttf"),
+        button: asset_server.load("images/main_menu/button.png"),
+
+        switch_base: asset_server.load("images/settings/switch_base.png"),
+        switch_head: asset_server.load("images/settings/switch_head.png"),
+
+        main_background: asset_server.load("images/main_menu/background.png"),
+        main_board: asset_server.load("images/main_menu/board.png"),
+        main_logo: asset_server.load("images/main_menu/bevypunk.png"),
+        settings_background: asset_server.load("images/settings/background.png"),
+    });
+}
